feat(changelog): add changelog components
Add the ability to configure sections of the changelog using a
`changelog_components` option.  Component outputs are separated by a blank
line and appear in the same order as they were configured.

It is possible to create your own custom components. Each component is a
function which returns either some text to be added, or None in which case it
will be skipped.

BREAKING CHANGE: The `compare_url` option has been removed in favor of using
`changelog_components`. This functionality is now available as the
`semantic_release.changelog.compare_url` component.
